+++ cmd/bnsd/app/init.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/iov-one/weave"
 	"github.com/iov-one/weave/app"
-	account "github.com/iov-one/weave/cmd/bnsd/x/account"
+	"github.com/iov-one/weave/cmd/bnsd/x/account"
 	"github.com/iov-one/weave/cmd/bnsd/x/username"
 	"github.com/iov-one/weave/coin"
 	"github.com/iov-one/weave/commands/server"
@@ -22,6 +22,7 @@ import (
 	"github.com/iov-one/weave/x/gov"
 	"github.com/iov-one/weave/x/msgfee"
 	"github.com/iov-one/weave/x/multisig"
+	"github.com/iov-one/weave/x/txfee"
 	"github.com/iov-one/weave/x/validators"
 	abci "github.com/tendermint/tendermint/abci/types"
 	"github.com/tendermint/tendermint/libs/log"
@@ -105,6 +106,7 @@ func DecorateApp(application app.BaseApp, logger log.Logger) app.BaseApp {
 		&gov.Initializer{},
 		&username.Initializer{},
 		&account.Initializer{},
+		&txfee.Initializer{},
 	))
 	application.WithLogger(logger)
 	return application