+++ x/txfee/doc.go
@@ -1,0 +1,8 @@
+/*
+Package txfee implements a fee middleware that computes an additional
+transaction fee, depending on each transaction binary size.
+
+Txfee extension is configured via `gconf` and price computation can be
+adjusted.
+*/
+package txfee